Significant rework of framework and function placement. Moved surface/cell creators to seperate file in utilities.
