+++ fridge/driver/assembly_holder.py
@@ -87,7 +87,6 @@ class Assembly:
         return:
             void
         """
-        print(plenum_data)
         self.plenum_smear_per = [plenum_data.ix['coolant_per', 'plenum'], plenum_data.ix['void_per', 'plenum'], plenum_data.ix['clad_per', 'plenum']]
         self.plenum_smear_zaids = [plenum_data.ix['coolant', 'plenum'], plenum_data.ix['void', 'plenum'], plenum_data.ix['clad', 'plenum']]
         self.plenum_material = []
@@ -95,10 +94,6 @@ class Assembly:
 
         self.fuel_reflector_smear_per = [fuel_reflector_data.ix['coolant_per', 'fuel_reflector'], fuel_reflector_data.ix['clad_per', 'fuel_reflector']]
         self.fuel_reflector_smear_zaids = [fuel_reflector_data.ix['coolant', 'fuel_reflector'], fuel_reflector_data.ix['clad', 'fuel_reflector']]
-
-        print(self.plenum_smear_zaids)
-        #plenum_material = [mat_read.material_reader([plenum_data.ix['plenum', 'fuel']])]
-        #fuel_reflecor_material = mat_read.material_reader([fuel_reflector_data.ix['bond', 'fuel']])
 
         # Potentially throw these three functions into the assembly holder and unpack them when they are needed.
         self.assembly_material = mat_read.material_reader([assembly_data.ix['assembly', 'assembly']])
@@ -112,7 +107,7 @@ class Assembly:
         self.surface_number = self.assembly_universe
         self.cell_number = self.assembly_universe + 50
         self.material_number = self.assembly_universe + 50
-        self.latice_universe = 0
+        self.lattice_universe = 0
 
         self.lower_reflector_surface = 0
         self.lower_reflector_mcnp_surface = ''
