+++ fridge/driver/surface_card_creator.py
@@ -1,0 +1,4 @@
+# import FRIDGe.fridge.driver.assembly_maker as am
+
+def mcnp_surface_card_creator(assembly):
+    return
